Add User Input to Load_Cell.ino
Instead of the weight/force being printed every 10 seconds, the program now asks users to input when they want to view the weight/force of the object. 
The purpose of this addition is to make the program more user friendly to new users who are unsure when to add the weight to the scale.